Create README.mdI
I was roaming around looking for where to get data to work with. I was particular about working on Excel to create visualization.

luckily and thankfully for me, I stumbled on Maven Analytics data playground. I was happy I did.

I saw a dataset on Coffee Shop Sales, so I jumped on it